+++ src/components/User/SystemAssets/components/AssetManager.vue
@@ -18,81 +18,112 @@
         <label class="asset-manager__image-lbl">Upload token logo</label>
         <image-field
           :fileKey="safeGet(asset, `details.logo.key`)"
-          @change="onFileChange($event, DOCUMENT_TYPES.tokenLogo)"/>
+          @change="onFileChange($event, DOCUMENT_TYPES.tokenLogo)"
+        />
       </div>
 
       <div class="app__form-row">
-        <input-field class="app__form-field"
+        <input-field
+          class="app__form-field"
           label="Asset name"
           v-model="asset.details.name"
           :disabled="isPending"
+          v-validate="'required|alpha'"
+          name="asset-name"
+          :errorMessage="errors.first('asset-name')"
         />
 
         <input-field class="app__form-field"
           label="Asset code"
           v-model="asset.code"
           :disabled="isExistingAsset || isPending"
+          v-validate="'required|alpha'"
+          name="asset-code"
+          :errorMessage="errors.first('asset-code')"
         />
       </div>
 
-      <!--TODO: need to add validation of public key-->
-
       <div class="app__form-row">
-        <input-field class="app__form-field"
+        <input-field
+          class="app__form-field"
           label="Issuer public key"
           v-model="asset.preissuedAssetSigner"
           :disabled="isExistingAsset || isPending"
+          v-validate="'required|alpha_num'"
+          name="issuer-key"
+          :errorMessage="errors.first('issuer-key')"
         />
 
-        <input-field class="app__form-field" v-if="!isExistingAsset"
+        <input-field
+          v-if="!isExistingAsset"
+          class="app__form-field"
           label="Initial preissued amount"
           v-model="asset.initialPreissuedAmount"
           :disabled="isExistingAsset || isPending"
+          v-validate="'required|decimal'"
+          name="initial-preissued"
+          :errorMessage="errors.first('initial-preissued')"
         />
 
-        <input-field v-else
-                     v-model="asset.availableForIssuance"
-                     class="app__form-field"
-                     label="Available for issuance"
-                     :disabled="true"
+        <input-field
+          v-else
+          v-model="asset.availableForIssuance"
+          class="app__form-field"
+          label="Available for issuance"
+          :disabled="true"
+          v-validate="'required|decimal'"
+          name="available-issuance"
+          :errorMessage="errors.first('available-issuance')"
         />
       </div>
 
       <div class="app__form-row">
-        <input-field class="app__form-field app__form-field--halved"
+        <input-field
+          class="app__form-field app__form-field--halved"
           type="number" :min="0" :step="DEFAULT_INPUT_STEP"
           label="Maximum tokens"
           v-model="asset.maxIssuanceAmount"
           :disabled="isExistingAsset || isPending"
+          v-validate="'required|decimal'"
+          name="max-tokens"
+          :errorMessage="errors.first('max-tokens')"
         />
       </div>
 
       <div class="asset-manager__file-input-wrp">
         <span>Terms</span>
         <div class="asset-manager__file-input-inner">
-          <label class="app__upload-btn app__btn app__btn--info"
-                 for="file-select">
+          <label
+            class="app__upload-btn app__btn app__btn--info"
+            for="file-select"
+          >
             Select File
           </label>
-          <input class="app__upload-input"
-                 id="file-select"
-                 type="file"
-                 accept="application/pdf, image/*"
-                 @change="onFileChange($event, DOCUMENT_TYPES.tokenTerms)"
+          <input
+            class="app__upload-input"
+            id="file-select"
+            type="file"
+            accept="application/pdf, image/*"
+            @change="onFileChange($event, DOCUMENT_TYPES.tokenTerms)"
           />
-          <span v-if="safeGet(asset, 'terms.name')" class="asset-manager__file-name">
-
+          <span
+            v-if="safeGet(asset, 'terms.name')"
+            class="asset-manager__file-name"
+          >
             {{ safeGet(asset, 'terms.name') }}
           </span>
           <span v-else-if="termsUrl" class="asset-manager__file-name">
-            <a :href="termsUrl" target="_blank" rel="noopener">{{ safeGet(asset, 'details.terms.name') }}</a>
+            <a :href="termsUrl" target="_blank" rel="noopener">
+              {{ safeGet(asset, 'details.terms.name') }}
+            </a>
           </span>
           <!--<span v-else-if="safeGet(asset, 'terms..')"></span>-->
         </div>
       </div>
 
       <div class="app__form-row">
-        <tick-field class="app__form-field"
+        <tick-field
+          class="app__form-field"
           v-model="asset.policy"
           :label="ASSET_POLICIES_VERBOSE[ASSET_POLICIES.transferable]"
           :cb-value="ASSET_POLICIES.transferable"
@@ -101,7 +132,8 @@
       </div>
 
       <div class="app__form-row">
-        <tick-field class="app__form-field"
+        <tick-field
+          class="app__form-field"
           v-model="asset.policy"
           :label="ASSET_POLICIES_VERBOSE[ASSET_POLICIES.baseAsset]"
           :cb-value="ASSET_POLICIES.baseAsset"
@@ -110,7 +142,8 @@
       </div>
 
       <div class="app__form-row">
-        <tick-field class="app__form-field"
+        <tick-field
+          class="app__form-field"
           v-model="asset.policy"
           :label="ASSET_POLICIES_VERBOSE[ASSET_POLICIES.statsQuoteAsset]"
           :cb-value="ASSET_POLICIES.statsQuoteAsset"
@@ -119,7 +152,8 @@
       </div>
 
       <div class="app__form-row">
-        <tick-field class="app__form-field"
+        <tick-field
+          class="app__form-field"
           v-model="asset.policy"
           :label="ASSET_POLICIES_VERBOSE[ASSET_POLICIES.withdrawable]"
           :cb-value="ASSET_POLICIES.withdrawable"
@@ -128,7 +162,8 @@
       </div>
 
       <div class="app__form-row">
-        <tick-field class="app__form-field"
+        <tick-field
+          class="app__form-field"
           v-model="asset.policy"
           :label="ASSET_POLICIES_VERBOSE[ASSET_POLICIES.withdrawableV2]"
           :cb-value="ASSET_POLICIES.withdrawableV2"
@@ -137,7 +172,8 @@
       </div>
 
       <div class="app__form-row">
-        <tick-field class="app__form-field"
+        <tick-field
+          class="app__form-field"
           v-model="asset.policy"
           :label="ASSET_POLICIES_VERBOSE[ASSET_POLICIES.twoStepWithdrawal]"
           title="Withdraw operations are done in two steps"
@@ -147,7 +183,8 @@
       </div>
 
       <div class="app__form-row">
-        <tick-field class="app__form-field"
+        <tick-field
+          class="app__form-field"
           v-model="asset.policy"
           :label="ASSET_POLICIES_VERBOSE[ASSET_POLICIES.requiresKyc]"
           title="Only users with KYC can submit/receive the asset"
@@ -157,7 +194,8 @@
       </div>
 
       <div class="app__form-row">
-        <tick-field class="app__form-field"
+        <tick-field
+          class="app__form-field"
           v-model="asset.policy"
           :label="ASSET_POLICIES_VERBOSE[ASSET_POLICIES.issuanceManualReviewRequired]"
           :cb-value="ASSET_POLICIES.issuanceManualReviewRequired"
@@ -168,8 +206,10 @@
       <div class="asset-manager-advanced__block">
         <div class="asset-manager-advanced__heading">
           <h3>Advanced</h3>
-          <button class="app__btn-secondary app__btn-secondary--iconed"
-                  @click.prevent="isShownAdvanced = !isShownAdvanced">
+          <button
+            class="app__btn-secondary app__btn-secondary--iconed"
+            @click.prevent="isShownAdvanced = !isShownAdvanced"
+          >
             <mdi-chevron-up-icon   v-if="isShownAdvanced"/>
             <mdi-chevron-down-icon v-else/>
           </button>
@@ -178,14 +218,15 @@
 
       <template v-if="isShownAdvanced">
         <div class="app__form-row">
-          <input-field class="app__form-field app__form-field--halved"
-                       type="number"
-                       label="External system type"
-                       name="External system type"
-                       v-model="asset.details.externalSystemType"
-                       :required="false"
-                       :disabled="isPending"
-                       v-validate="{max_value: int32}"
+          <input-field
+            class="app__form-field app__form-field--halved"
+            type="number"
+            label="External system type"
+            name="External system type"
+            v-model="asset.details.externalSystemType"
+            :required="false"
+            :disabled="isPending"
+            v-validate="{max_value: int32}"
           />
         </div>
       </template>
@@ -287,6 +328,7 @@ export default {
     },
 
     async submit () {
+      if (!this.isValid()) return
       if (!await confirmAction()) return
 
       this.isPending = true
@@ -301,7 +343,7 @@ export default {
             requestID: '0',
             code: String(this.asset.code),
             policies: Number(this.asset.policy),
-            logoId: String(this.asset.logoId || this.asset.logoId),
+            logoId: String(this.asset.logoId),
             details: {
               name: this.asset.details.name,
               externalSystemType: this.asset.details.externalSystemType
@@ -311,8 +353,8 @@ export default {
           operation = Sdk.base.ManageAssetBuilder.assetCreationRequest({
             requestID: '0',
             code: String(this.asset.code),
-            preissuedAssetSigner: String(this.asset.preissuedAssetSigner || this.asset.preissuedAssetSigner),
-            maxIssuanceAmount: String(this.asset.maxIssuanceAmount || this.asset.maxIssuanceAmount),
+            preissuedAssetSigner: String(this.asset.preissuedAssetSigner),
+            maxIssuanceAmount: String(this.asset.maxIssuanceAmount),
             policies: Number(this.asset.policy),
             initialPreissuedAmount: this.asset.initialPreissuedAmount,
             details: {
@@ -329,6 +371,18 @@ export default {
         error.showMessage()
       }
       this.isPending = false
+    },
+
+    isValid () {
+      if (this.errors.errors.lenght) {
+        this.$store.dispatch('SET_ERROR', 'Some field is invalid')
+        return false
+      }
+      if (!Sdk.base.Keypair.isValidPublicKey(this.asset.preissuedAssetSigner)) {
+        this.$store.dispatch('SET_ERROR', 'Issuer public key is invalid')
+        return false
+      }
+      return true
     },
 
     async onFileChange (event, type) {